fix: add non inline exmaple and use function instead of task for running server
