+++ examples/deploy-example/contracts/ChildChildLib.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.11;
+library ChildChildLib {
+
+    function plus(uint a, uint b) public pure returns (uint) {
+        return a - b;
+    }
+}
