+++ packages/hardhat-zksync-node/src/index.ts
@@ -11,7 +11,6 @@ import { HARDHAT_NETWORK_NAME } from 'hardhat/plugins';
 import {
     MAX_PORT_ATTEMPTS,
     START_PORT,
-    TASK_NODE_GET_SERVER,
     TASK_NODE_ZKSYNC,
     TASK_NODE_ZKSYNC_CREATE_SERVER,
     TASK_NODE_ZKSYNC_DOWNLOAD_BINARY,
@@ -292,24 +291,5 @@ task(
         }
     },
 );
-
-task(TASK_NODE_GET_SERVER, async (_args, { run }) => {
-    const platform = getPlatform();
-    if (platform === 'windows' || platform === '') {
-        throw new ZkSyncNodePluginError(`Unsupported platform: ${platform}`);
-    }
-
-    // Download the binary, if necessary
-    const binaryPath: string = await run(TASK_NODE_ZKSYNC_DOWNLOAD_BINARY, { force: false });
-
-    const currentPort = await getAvailablePort(START_PORT, MAX_PORT_ATTEMPTS);
-    const commandArgs = constructCommandArgs({ port: currentPort });
-
-    return {
-        commandArgs,
-        server: new JsonRpcServer(binaryPath),
-        port: currentPort,
-    };
-});
 
 export { ZkSyncProviderAdapter } from './zksync-provider-adapter';
