+++ packages/hardhat-zksync-node/src/utils.ts
@@ -8,6 +8,7 @@ import { exec } from 'child_process';
 import type { Dispatcher } from 'undici';
 import { Provider } from 'zksync-ethers';
 import { getCompilersDir } from 'hardhat/internal/util/global-dir';
+import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { ZkSyncProviderAdapter } from './zksync-provider-adapter';
 
 import {
@@ -19,16 +20,21 @@ import {
     ALLOWED_SHOW_STORAGE_LOGS_VALUES,
     ALLOWED_SHOW_VM_DETAILS_VALUES,
     BASE_URL,
+    MAX_PORT_ATTEMPTS,
     NETWORK_ACCOUNTS,
     NETWORK_ETH,
     NETWORK_GAS,
     NETWORK_GAS_PRICE,
     PLATFORM_MAP,
+    START_PORT,
+    TASK_NODE_ZKSYNC_DOWNLOAD_BINARY,
     TEMP_FILE_PREFIX,
     ZKSYNC_ERA_TEST_NODE_NETWORK_NAME,
 } from './constants';
 import { ZkSyncNodePluginError } from './errors';
 import { CommandArguments } from './types';
+import { JsonRpcServer } from './server';
+import './index';
 
 // Generates command arguments for running the era-test-node binary
 export function constructCommandArgs(args: CommandArguments): string[] {
@@ -405,3 +411,21 @@ export function configureNetwork(network: any, port: number) {
     network.config = getNetworkConfig(url);
     network.provider = new ZkSyncProviderAdapter(new Provider(url));
 }
+
+export const startServer = async (hre: HardhatRuntimeEnvironment) => {
+    const platform = getPlatform();
+    if (platform === 'windows' || platform === '') {
+        throw new ZkSyncNodePluginError(`Unsupported platform: ${platform}`);
+    }
+
+    const binaryPath: string = await hre.run(TASK_NODE_ZKSYNC_DOWNLOAD_BINARY, { force: false });
+
+    const currentPort = await getAvailablePort(START_PORT, MAX_PORT_ATTEMPTS);
+    const commandArgs = constructCommandArgs({ port: currentPort });
+
+    return {
+        commandArgs,
+        server: new JsonRpcServer(binaryPath),
+        port: currentPort,
+    };
+};
